requirements.txt
pandas==2.1.2
numpy==1.26.2
scipy==1.11.3
scikit-learn==1.3.0
reportlab==4.0.0
joblib==1.3.2
python-dateutil==2.8.2
pyyaml==6.0
